Deduplicate the Dag run lookup in airflowctl dag and task commands
The dags and tasks commands each carried their own copy of the same
logical-date Dag run lookup, differing only in what they returned, and each
repeated the run-selector guard in front of it. Sharing them lets the selector
handling live in one place rather than in front of every command that accepts
one.

A supplied run_id is deliberately still taken at face value rather than
fetched, so commands acting on a nested resource keep reporting a miss against
that resource instead of against the Dag run. The added assertions pin that
down, since collapsing the two resolvers would otherwise silently add a
request and re-attribute the 404.
