+++ airflow-ctl/src/airflowctl/ctl/commands/dag_command.py
@@ -27,7 +27,6 @@ from rich.text import Text
 
 from airflowctl.api.client import (
     NEW_API_CLIENT,
-    Client,
     ClientKind,
     ServerResponseError,
     provide_api_client,
@@ -35,9 +34,9 @@ from airflowctl.api.client import (
 from airflowctl.api.datamodels.generated import (
     ClearTaskInstancesBody,
     DAGPatchBody,
-    DAGRunResponse,
 )
 from airflowctl.ctl.console_formatting import AirflowConsole
+from airflowctl.ctl.utils.dag_run import resolve_dag_run
 
 
 def update_dag_state(
@@ -118,58 +117,10 @@ def next_execution(args, api_client=NEW_API_CLIENT) -> dict | None:
     return result
 
 
-def _get_dag_run_by_run_id(api_client: Client, dag_id: str, run_id: str) -> DAGRunResponse:
-    """Get a Dag run by its run ID."""
-    try:
-        return api_client.dag_runs.get(dag_id=dag_id, dag_run_id=run_id, suppress_error_log=True)
-    except ServerResponseError as e:
-        if e.response.status_code != 404:
-            raise
-        rich.print(f"[red]Dag run {run_id!r} of Dag {dag_id!r} not found[/red]")
-        sys.exit(1)
-
-
-def _get_dag_run_by_logical_date(api_client: Client, dag_id: str, value: str) -> DAGRunResponse:
-    """Get the Dag run with an exact logical date match."""
-    try:
-        logical_date = datetime.datetime.fromisoformat(value.replace("Z", "+00:00"))
-    except ValueError:
-        rich.print(f"[red]Invalid --logical-date: {value!r}[/red]")
-        sys.exit(1)
-    if logical_date.tzinfo is None:
-        rich.print("[red]--logical-date must include a timezone offset[/red]")
-        sys.exit(1)
-
-    dag_runs = []
-    try:
-        dag_runs = api_client.dag_runs.list(
-            dag_id=dag_id,
-            logical_date_gte=logical_date,
-            logical_date_lte=logical_date,
-            order_by="-id",
-            limit=1,
-            suppress_error_log=True,
-        ).dag_runs
-    except ServerResponseError as e:
-        if e.response.status_code != 404:
-            raise
-    if not dag_runs:
-        rich.print(f"[red]Dag run for {dag_id} with logical date {value!r} not found[/red]")
-        sys.exit(1)
-    return dag_runs[0]
-
-
 @provide_api_client(kind=ClientKind.CLI)
 def state(args, api_client=NEW_API_CLIENT) -> None:
     """Show the state and configuration of a Dag run."""
-    if (args.run_id is None) == (args.logical_date is None):
-        rich.print("[red]Provide either run_id or --logical-date, but not both[/red]")
-        sys.exit(1)
-
-    if args.run_id:
-        dag_run = _get_dag_run_by_run_id(api_client, args.dag_id, args.run_id)
-    else:
-        dag_run = _get_dag_run_by_logical_date(api_client, args.dag_id, args.logical_date)
+    dag_run = resolve_dag_run(api_client, args)
 
     state_value = getattr(dag_run.state, "value", dag_run.state)
     if dag_run.conf:
