+++ airflow-ctl/src/airflowctl/ctl/commands/task_command.py
@@ -17,7 +17,6 @@
 
 from __future__ import annotations
 
-import datetime
 import sys
 from typing import TYPE_CHECKING
 
@@ -26,39 +25,10 @@ import rich
 from airflowctl.api.client import NEW_API_CLIENT, ClientKind, ServerResponseError, provide_api_client
 from airflowctl.api.datamodels.generated import TaskInstanceState
 from airflowctl.ctl.console_formatting import AirflowConsole
+from airflowctl.ctl.utils.dag_run import resolve_dag_run_id
 
 if TYPE_CHECKING:
     from airflowctl.api.datamodels.generated import TaskInstanceResponse
-
-
-def _find_run_id_by_logical_date(api_client, dag_id: str, value: str) -> str:
-    """Find the run ID of the Dag run with an exact logical date match."""
-    try:
-        logical_date = datetime.datetime.fromisoformat(value.replace("Z", "+00:00"))
-    except ValueError:
-        rich.print(f"[red]Invalid --logical-date: {value!r}[/red]")
-        sys.exit(1)
-    if logical_date.tzinfo is None:
-        rich.print("[red]--logical-date must include a timezone offset[/red]")
-        sys.exit(1)
-
-    dag_runs = []
-    try:
-        dag_runs = api_client.dag_runs.list(
-            dag_id=dag_id,
-            logical_date_gte=logical_date,
-            logical_date_lte=logical_date,
-            order_by="-id",
-            limit=1,
-            suppress_error_log=True,
-        ).dag_runs
-    except ServerResponseError as e:
-        if e.response.status_code != 404:
-            raise
-    if not dag_runs:
-        rich.print(f"[red]Dag run for {dag_id} with logical date {value!r} not found[/red]")
-        sys.exit(1)
-    return dag_runs[0].dag_run_id
 
 
 def _format_task_instance(ti: TaskInstanceResponse, has_mapped_instances: bool) -> dict[str, str]:
@@ -78,11 +48,7 @@ def _format_task_instance(ti: TaskInstanceResponse, has_mapped_instances: bool) 
 @provide_api_client(kind=ClientKind.CLI)
 def failed_deps(args, api_client=NEW_API_CLIENT) -> None:
     """Get task instance dependencies that were not met, from the scheduler's perspective."""
-    if (args.run_id is None) == (args.logical_date is None):
-        rich.print("[red]Provide either run_id or --logical-date, but not both[/red]")
-        sys.exit(1)
-
-    run_id = args.run_id or _find_run_id_by_logical_date(api_client, args.dag_id, args.logical_date)
+    run_id = resolve_dag_run_id(api_client, args)
 
     try:
         response = api_client.task_instances.get_dependencies(
@@ -129,11 +95,7 @@ def failed_deps(args, api_client=NEW_API_CLIENT) -> None:
 @provide_api_client(kind=ClientKind.CLI)
 def states_for_dag_run(args, api_client=NEW_API_CLIENT) -> None:
     """Get the status of all task instances in a Dag run."""
-    if (args.run_id is None) == (args.logical_date is None):
-        rich.print("[red]Provide either run_id or --logical-date, but not both[/red]")
-        sys.exit(1)
-
-    run_id = args.run_id or _find_run_id_by_logical_date(api_client, args.dag_id, args.logical_date)
+    run_id = resolve_dag_run_id(api_client, args)
 
     try:
         task_instances = api_client.task_instances.list(dag_id=args.dag_id, dag_run_id=run_id).task_instances
